moved SSOAgentConstants to util package
