+++ components/org.wso2.carbon.identity.sso.agent/src/main/java/org/wso2/carbon/identity/sso/agent/oauth2/SAML2GrantManager.java
@@ -22,7 +22,7 @@ package org.wso2.carbon.identity.sso.agent.oauth2;
 
 import com.google.gson.Gson;
 import org.opensaml.xml.util.Base64;
-import org.wso2.carbon.identity.sso.agent.SSOAgentConstants;
+import org.wso2.carbon.identity.sso.agent.util.SSOAgentConstants;
 import org.wso2.carbon.identity.sso.agent.exception.SSOAgentException;
 import org.wso2.carbon.identity.sso.agent.bean.LoggedInSessionBean;
 import org.wso2.carbon.identity.sso.agent.bean.SSOAgentConfig;
